fix: linting errors in ErrorMarkerWithMessage class
